GEODE-3254: Refactoring ConfigCommands
